+++ src/Pages/HomePage/index.jsx
@@ -1,20 +1,12 @@
-import React, { useState, useEffect, Profiler } from "react";
+import React, { useState, useEffect } from "react";
 import axios from "axios";
-import Typography from "@mui/material/Typography";
 import AttendanceTable from "./AttendanceTable";
 import AttendanceButton from "./AttendanceButton";
-import SetTodayWordButton from "./SetSetTodayWordButton";
+import SetTodayWordButton from "./SetTodayWordButton";
+import UserProfile from "./UserProfile";
+import TodayDate from "./TodayDate";
 import TestButtons from "./TestButtons";
 import useStore from "../../store.js";
-import { format } from "date-fns";
-import { ko } from "date-fns/locale";
-import UserProfile from "./UserProfile";
-
-// 오늘 날짜를 "yyyy년 MM월 dd일 e요일" 형식으로 변환
-const getTodayDate = () => {
-  const todayDate = new Date();
-  return `${format(todayDate, "PPP EEEE", { locale: ko })}`;
-};
 
 const Home = () => {
   const { _intraId, _photoUrl } = useStore((state) => state);
@@ -55,9 +47,7 @@ const Home = () => {
   return (
     <>
       <UserProfile intraId={_intraId} photoUrl={_photoUrl} />
-      <Typography variant="subtitle1" sx={{ mt: 1, fontWeight: "bold" }}>
-        {getTodayDate()}
-      </Typography>
+      <TodayDate />
       <AttendanceTable summary={summary} />
       <AttendanceButton />
       {isOperator && <SetTodayWordButton />}
